+++ CHANGELOG.md
@@ -6,11 +6,17 @@
  *   Debug shows the position and name of every layer
  *   Allow variables and expressions in parameters from API
  *   Fix: Check that = has a value
+ *   Fix: Default degrees value in arc() is now undefined
+ *   Fix: Added parameter tips for =
  *   Fix: Inorder instructions don't show null first param when adding floating block
  *   Fix: In guide do not show "undefined" in steps titles if no title has been defined
  *   Fix: In Block setup the tip about parameters only shows () if the instruction uses brackets
  *   Fix: Accept read correctly variables and expressions in number fields in setupBlock
  *   Fix: Allow steps with no message in guide
+ *   Fix: arc() icon
+ *   Fix: encodeURI/encodeURIComponent mess in guide
+ *   Fix: Reset code when going back to first steps in guide
+ *   Fix: Custom instructions in guide are now reset when reloaded
 
 ## 2.3 (2015-12-06)